grabber_service/requirements.txt
Pyrogram
configparser
PyMySQL
cryptography
pycryptodome
TgCrypto
